Finish type annotations for .parsing
This also adds scanner support for "callback"
functions/blocks where one of the arguments is a function or block.
Supports one level of nesting, I might add more levels later as needed.

The code for this is suboptimal due to a limitation in mypy: that tool
does not yet support cyclic definitions in types.