Partial implementation for a document generator
The goal for this is to emit a sphinx document for a
framework based on the collected metadata. Note that
this will never be a lot better than a raw dump of
the package contents (like 'help(AppKit)'), although
I do want to merge the "apinotes" documents into the
generated documentation.

This initial checkin only documents enum and extern
definitions.