Improve generated documentation.
The output file turns out to be very large and likely needs to
be split to be somewhat usable.

This commit adds documentation for classes and their methods,
albeith still incomplete.

As mentioned in earlier commits the documentation is just mechanical,
and hence doesn't mention what methods actually do, but do describe
their interface.