Coax the parser into parsing the Security framework headers
The encoding for struct types is not quite to my liking yet, but
with some luck that's for types that I don't want to wrap in the
first place (and otherwise is should be possible to fix the issue)

The parsing code could do with documentation and cleanup, it is
a bit too magical. But that can wait, first get the tooling in a
good enough shape to be able to create compiled metadata files again.