Convert .parsing to the new datamodel
The pre-calculated exceptions data is empty for now,
will be fixed later.

Also adds full type annotations, which required adding
a lot of asserts. It should be possible to clean this up
by improving the type annotation with some code cleanup.

I also have full type checking for .clang, but have broken
that module in the proces so this will have to wait.